fix(plugin): use an empty frontchannel logout URL instead of omitting it
This commit adds a check to send an empty frontchannel logout URL to the
frontend if no URL is set. This check already exists for the backchannel
logout URL and now behaves identically.
With that, it fixes an unintentional error message in the frontend when
removing focus from an empty frontchannel logout URL field.

Fixes #11