refactor: remove client limit for users
This commit removes the limit for self service clients per user.
The reason for that is that checking the client limit is harder to
check correctly with our simple model of client sharing.

Keycloak itself has many ways to abuse the internal resources and there
seems to be no limiting mechanism implemented in Keycloak.
Implementing the limit in the plugin seems like the wrong layer for that
feature and we could come back to this, once Keycloak has an internal
mechanism.